+++ src/main/resources/db/migration/V9__create_person2dialog.sql
@@ -1,5 +1,5 @@
 CREATE TABLE person2dialog (
 	id bigserial not null,
-    person_id bigserial not null,
-    dialog_id bigserial not null,
+    person_id int8 not null,
+    dialog_id int8 not null,
     primary key (id));
